+++ backend/src/message/message.module.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { MessageGateway } from "./message.gateway";
+
+@Module({
+  imports: [],
+  providers: [MessageGateway],
+})
+export class MessageModule {}