+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ## Features:
 
 Backend:
+1. Deleted alerts
+2. Added message module, model
+3. Changed User model
+
+# v2.5.1 (16.03.2022)
+
+## Features:
+
+Backend:
 1. Added alert namespace
 2. Updated chat namespace